% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{envir.prep}
\alias{envir.prep}
\title{Library initialize and path loader}
\usage{
envir.prep()
}
\value{
library location
}
\description{
Initialize and load library path for package
}
\examples{
envir.prep()
}
