% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{equationSymbol.Widget}
\alias{equationSymbol.Widget}
\title{Primary widget container}
\usage{
equationSymbol.Widget(viewer = c("pane", "dialog"))
}
\arguments{
\item{viewer}{viewer option to open widget}
}
\value{
shiny app widget
}
\description{
Customizable widget opener
}
\examples{
if(interactive()){
equationSymbol.Widget()
}

}
