% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{insert.copied}
\alias{insert.copied}
\title{Paste items from clipboard}
\usage{
insert.copied()
}
\value{
pastes copied items into current cursor position
}
\description{
Insert copied item from clipboard
}
\examples{
insert.copied()
}
