% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{insert.Symbol.Raw}
\alias{insert.Symbol.Raw}
\title{Add symbol to current page}
\usage{
insert.Symbol.Raw(symbolid, type = letters[1:4])
}
\arguments{
\item{symbolid}{the symbol id to include}

\item{type}{type of inclusion}
}
\value{
pastes copied symbol code into document
}
\description{
Add selected symbol to page
}
\examples{
if(interactive()){
insert.Symbol.Raw("sample copied")
insert.Symbol.Raw("rpkg.net","a")
}
}
