% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{r2sym.server}
\alias{r2sym.server}
\title{User Interface Server}
\usage{
r2sym.server(input, output, session)
}
\arguments{
\item{input}{function input}

\item{output}{function output}

\item{session}{app session}
}
\value{
server for user interface to control insertions
}
\description{
Assembles Server function for the user interface
}
\examples{
if(interactive()){
library(shiny)
ui <- fluidPage(
  "Sample rpkg.net"
)

server <- symbol.equation.gpt::r2sym.server

shinyApp(ui, server)
}

}
