% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{r2sym.ui}
\alias{r2sym.ui}
\title{App User Interface}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
r2sym.ui
}
\value{
HTML for user interface container for the builder
}
\description{
Assembles UI for the application
}
\examples{
r2sym.ui
}
\keyword{datasets}
