% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{r2sym.ui2}
\alias{r2sym.ui2}
\title{UI}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 1.
}
\usage{
r2sym.ui2
}
\description{
Assembles UI
}
\keyword{datasets}
