% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/symbol.equation.ui-package.R
\docType{package}
\name{symbol.equation.ui}
\alias{symbol.equation.ui}
\alias{symbol.equation.ui-package}
\title{Main Widget Helper for symbols}
\usage{
symbol.equation.ui()
}
\value{
user interface for the builder
}
\description{
Assembles the interface for the symbols builder

User Interface for adding symbols, smileys, arrows, building equations using 'LaTeX' or 'r2symbols'. Built for use in development of 'Markdown' and 'Shiny' Outputs.
}
\examples{
symbol.equation.ui()
}
\seealso{
Useful links:
\itemize{
  \item \url{https://symbols-ui.obi.obianom.com/}
  \item Report bugs at \url{https://github.com/oobianom/r2symbols.ui/issues/}
}

}
\author{
\strong{Maintainer}: Obinna Obianom \email{idonshayo@gmail.com}

}
\keyword{internal}
