% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{symbol.equation.UI2}
\alias{symbol.equation.UI2}
\title{Main Widget Helper for symbols viewer2}
\usage{
symbol.equation.UI2()
}
\value{
user interface for the builder in modal
}
\description{
Assembles the interface for the symbols builder
}
\examples{
symbol.equation.UI2()
}
