% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{insert.Symbol.Raw}
\alias{insert.Symbol.Raw}
\title{Add symbol to current page}
\usage{
insert.Symbol.Raw(symbolid, type = letters[1:4])
}
\arguments{
\item{symbolid}{the symbol id to include}

\item{type}{type of inclusion}
}
\description{
Add selected symbol to page
}
