% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{r2sym.ui}
\alias{r2sym.ui}
\title{App User Interface}
\usage{
r2sym.ui()
}
\value{
HTML for user interface container for the builder
}
\description{
Assembles UI for the application
}
\examples{
\dontrun{
if(interactive()){
r2sym.ui()
}
}
}
