% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.equation.ui-package.R
\docType{package}
\name{symbol.equation.ui}
\alias{symbol.equation.ui}
\alias{symbol.equation.ui-package}
\title{symbol.equation.ui}
\description{
User Interface for adding symbols, smileys, arrows, building equations using 'LaTeX' or 'r2symbols'. Built for use in development of 'Markdown' and 'Shiny' Outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://symbols-ui.obi.obianom.com/}
  \item Report bugs at \url{https://github.com/oobianom/r2symbols.ui/issues/}
}

}
\author{
\strong{Maintainer}: Obinna Obianom \email{idonshayo@gmail.com}

}
\keyword{internal}
