% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{symbol.equation.ui}
\alias{symbol.equation.ui}
\title{Main Widget Helper for symbols}
\usage{
symbol.equation.ui()
}
\value{
user interface for the builder
}
\description{
Assembles the interface for the symbols builder
}
\examples{
\dontrun{
symbol.equation.ui()
}
}
